Implement logic for SearchViewModel
