+++ UberGoCore/UberGoCore/MapViewModel.swift
@@ -22,7 +22,6 @@ public protocol MapViewModelProtocol {
 public protocol MapViewModelInput {
 
     var startUpdateLocationTriggerPublisher: PublishSubject<Bool> { get }
-    var textSearchPublish: PublishSubject<String> { get }
     var didSelectPlaceObjPublisher: PublishSubject<PlaceObj?> { get }
     var routeToDestinationPublisher: PublishSubject<PlaceObj> { get }
     var routeForCurrentTripPublisher: PublishSubject<TripObj> { get }
@@ -32,10 +31,6 @@ public protocol MapViewModelOutput {
 
     // Origin
     var currentPlaceDriver: Driver<PlaceObj> { get }
-
-    // Search
-    var searchPlaceObjsVariable: Variable<[PlaceObj]> { get }
-    var loadingDriver: Driver<Bool> { get }
 
     // Destination
     var selectedPlaceObjDriver: Driver<PlaceObj?> { get }
@@ -59,38 +54,34 @@ open class MapViewModel:
     public var output: MapViewModelOutput { return self }
 
     // MARK: - Variable
-    fileprivate let mapManager: MapService
+    fileprivate let mapService: MapService
     fileprivate let uberService: UberService
     fileprivate let directionService: DirectionService
 
     // MARK: - Input
     public var startUpdateLocationTriggerPublisher = PublishSubject<Bool>()
-    public var textSearchPublish = PublishSubject<String>()
     public var didSelectPlaceObjPublisher = PublishSubject<PlaceObj?>()
     public var routeForCurrentTripPublisher = PublishSubject<TripObj>()
     public var routeToDestinationPublisher = PublishSubject<PlaceObj>()
 
     // MARK: - Output
-    public var currentLocationVariable: Variable<CLLocation?> { return mapManager.output.currentLocationVar }
+    public var currentLocationVariable: Variable<CLLocation?> { return mapService.output.currentLocationVar }
     public var currentPlaceDriver: Driver<PlaceObj> {
-        return mapManager.output.currentPlaceObs
+        return mapService.output.currentPlaceObs
             .asDriver(onErrorJustReturn: PlaceObj.invalid)
     }
-    public var searchPlaceObjsVariable = Variable<[PlaceObj]>([])
-    fileprivate var personalOrHistoryPlaceObjsVariable = Variable<[PlaceObj]>([])
-    public let loadingDriver: Driver<Bool>
     public var selectedPlaceObjDriver: Driver<PlaceObj?>
     public var selectedDirectionRouteObserver: Observable<Route?>
     public var isSelectedPlace: Driver<Bool> { return selectedPlaceObjDriver.map({ $0 != nil }) }
     public var routeCurrentTrip: Driver<Route?>
 
     // MARK: - Init
-    public init(mapManager: MapService = MapService(),
+    public init(mapService: MapService,
                 uberService: UberService = UberService(),
                 directionService: DirectionService = DirectionService(),
                 googleMapService: GoogleMapService = GoogleMapService()) {
 
-        self.mapManager = mapManager
+        self.mapService = mapService
         self.uberService = uberService
         self.directionService = directionService
 
@@ -100,80 +91,12 @@ open class MapViewModel:
             .distinctUntilChanged()
             .subscribe(onNext: {(trigger) in
                 if trigger {
-                    mapManager.startUpdatingLocation()
+                    mapService.startUpdatingLocation()
                 } else {
-                    mapManager.stopUpdatingLocation()
+                    mapService.stopUpdatingLocation()
                 }
             })
             .addDisposableTo(disposeBag)
-
-        // Load personal or history place
-        let personalPlace = uberService
-            .personalPlaceObserver()
-            .startWith([]) // Don't need to wait -> Should show history palce first
-
-        let historyPlace = uberService.historyObserver
-
-        Observable.combineLatest([personalPlace, historyPlace])
-        .map { (combine) -> [PlaceObj] in
-            let personPlaces = combine.first!
-            let historyPlaces = combine.last!
-            return personPlaces + historyPlaces
-        }.bind(to: personalOrHistoryPlaceObjsVariable)
-        .addDisposableTo(disposeBag)
-
-        // Trigger
-        uberService.reloadHistoryTrigger.onNext()
-
-        // Text Search
-        let shared = textSearchPublish
-            .asObservable()
-            .share()
-
-        // Reset Empty data while fetching
-        let emptyOb = shared
-            .filter({ $0 != "" })
-            .map { _ -> [PlaceObj] in
-                return []
-            }
-
-        // Search from Google Service
-        let searchPlaceOb = shared
-            .filter { $0 != "" }
-            .debounce(0.3, scheduler: MainScheduler.instance)
-            .filter { $0 != "" }
-            .distinctUntilChanged()
-            .withLatestFrom(mapManager.output.currentLocationVar.asObservable().filterNil(),
-                            resultSelector: { (namePlace, location) -> (String, CLLocationCoordinate2D) in
-                                return (namePlace, location.coordinate)
-            })
-            .flatMapLatest { return googleMapService.searchPlaces(with: $0.0, currentLocation: $0.1) }
-            .share()
-
-        let personalOrHistoryOb = shared
-            .distinctUntilChanged()
-            .filter({ $0 == "" })
-            .withLatestFrom(personalOrHistoryPlaceObjsVariable.asObservable())
-            .share()
-
-        // Merage into searchPlace
-        let searchFinishOb = Observable.merge([searchPlaceOb,
-                          personalOrHistoryOb,
-                          emptyOb,
-                          personalOrHistoryPlaceObjsVariable.asObservable()])
-            .skip(1)
-            .share()
-
-        searchFinishOb
-            .bind(to: searchPlaceObjsVariable)
-            .addDisposableTo(disposeBag)
-
-        // Loader
-        let startLoadingOb = Observable.merge([searchPlaceOb, personalOrHistoryOb]).map { _ in true }
-        let stopLoadingOb = searchFinishOb.map { _ in false }
-        loadingDriver = Observable.merge([startLoadingOb, stopLoadingOb])
-            .map({ !$0 })
-            .asDriver(onErrorJustReturn: false)
 
         // Selected
         let selectedPlaceObserve = didSelectPlaceObjPublisher
@@ -185,7 +108,7 @@ open class MapViewModel:
 
         // Get Route
         let currentPlaceObj = routeToDestinationPublisher
-            .withLatestFrom(mapManager.currentPlaceObs.asObservable())
+            .withLatestFrom(mapService.currentPlaceObs.asObservable())
         let getDirection = Observable.zip([selectedPlaceObserve.filterNil(), currentPlaceObj])
             .flatMapLatest { data -> Observable<Route?> in
                 let from = data.first!
