--- UberGoCore/UberGoCore/SearchBarViewModel.swift
@@ -1,16 +1,0 @@
-//
-//  SearchBarViewModel.swift
-//  UberGoCore
-//
-//  Created by Nghia Tran on 6/5/17.
-//  Copyright © 2017 Nghia Tran. All rights reserved.
-//
-
-import Cocoa
-import RxCocoa
-import RxSwift
-
-// MARK: - SearchBarViewModel
-open class SearchBarViewModel: BaseViewModel {
-
-}
